# infra/APIContracts/src/main/resources/adverts-crud-openapi.yaml
openapi: 3.1.0

info:
  title: Модуль приложения eBoard реализующий CRUD-операции для сущности Advert
  version: 0.0.3

servers:
  - url: http://localhost:8082
    description: адрес сервера
  
paths:
  /adverts/new:
    post:
      operationId: createAdvert
      tags:
        - Advert
      description: Создать новое объявление
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertDTO"
      responses:
        '200':
          description: OK
        'default':
          description: Что-то пошло не так
  /adverts/{uuid}:
    get:
      operationId: getAdvert
      tags:
        - Advert
      description: Получить данные о объявлении
      parameters:
        - $ref: "common.yaml#/components/parameters/uuid"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvertDTO"
    put:
      operationId: updateAdvert
      tags:
        - Advert
      description: Обновить данные о объявлении
      parameters:
        - $ref: "common.yaml#/components/parameters/uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertDTO"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvertDTO"
    delete:
      operationId: deleteAdvert
      tags:
        - Advert
      description: Удалить данные о объявлении
      parameters:
        - $ref: "common.yaml#/components/parameters/uuid"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvertDTO"

components:
  schemas:
    AdvertDTO:
      title: Объявление
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID объявления
        title:
          type: string
          description: Заголовок объявления
        cost:
          type: number
          format: float
        description:
          type: string
          description: Описание объявления
        owner:
          type: object
          schema:
            $ref: 'users-crud-openapi.yaml#/components/schemas/UserDTO'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []